Move TypePtr out of Types.h
