+++ core/Types.h
@@ -5,10 +5,10 @@
 #include "core/Context.h"
 #include "core/Error.h"
 #include "core/SymbolRef.h"
+#include "core/TypeConstraint.h"
 #include <memory>
 #include <optional>
 #include <string>
-
 namespace sorbet::core {
 /** Dmitry: unlike in Dotty, those types are always dealiased. For now */
 class Type;
@@ -23,49 +23,6 @@ class Symbol;
 class TypeVar;
 class SendAndBlockLink;
 class TypeAndOrigins;
-// using TypePtr = std::shared_ptr<Type>;
-
-class TypePtr {
-    std::shared_ptr<Type> store;
-    TypePtr(std::shared_ptr<Type> &&store);
-
-public:
-    TypePtr() = default;
-    TypePtr(TypePtr &&other) = default;
-    TypePtr(const TypePtr &other) = default;
-    TypePtr &operator=(TypePtr &&other) = default;
-    TypePtr &operator=(const TypePtr &other) = default;
-    explicit TypePtr(Type *ptr) : store(ptr) {}
-    TypePtr(std::nullptr_t n) : store(nullptr) {}
-    operator bool() const {
-        return (bool)store;
-    }
-    Type *get() const {
-        return store.get();
-    }
-    Type *operator->() const {
-        return get();
-    }
-    Type &operator*() const {
-        return *get();
-    }
-    bool operator!=(const TypePtr &other) const {
-        return store != other.store;
-    }
-    bool operator==(const TypePtr &other) const {
-        return store == other.store;
-    }
-    bool operator!=(std::nullptr_t n) const {
-        return store != nullptr;
-    }
-    bool operator==(std::nullptr_t n) const {
-        return store == nullptr;
-    }
-    friend class Symbol;
-
-    template <class T, class... Args> friend TypePtr make_type(Args &&... args);
-};
-CheckSize(TypePtr, 16, 8);
 
 class ArgInfo {
 public:
